restructuring and optimizing
