+++ backend/app/app/api/api_v1/endpoints/audio.py
@@ -1,10 +1,11 @@
 from fastapi import APIRouter, Depends
-from ....crud import crud_audio
-from ....schemas import audio
 from ... import deps
 from sqlalchemy.orm import Session
-from ....crud import s3
+from ....crud.s3 import s3_audio
+from ....schemas import audio
 from ....models import models
+from typing import List
+from ....crud import crud_audio
 
 
 
@@ -17,14 +18,17 @@ router = APIRouter(
 @router.post("/text_to_audio/{book_id}")
 def text_to_audio(book_id: int, file_name: str, db: Session = Depends(deps.get_db)):
     # Post audio streams
-    s3.save_audio_to_s3(book_id=book_id, file_name=file_name, db=db)
+    s3_audio.save_audio_to_s3(book_id=book_id, file_name=file_name, db=db)
     return 'Saved successfully'
+
+# works
+@router.get("/", response_model=List[audio.Audio])
+def read_Audio(skip: int = 0, limit: int = 100, db: Session = Depends(deps.get_db)):
+    audio = crud_audio.Audio.get_multi_Audios(db, skip=skip, limit=limit)
+    return audio
 
 @router.delete("/{audio_id}")
 def delete_audio(audio_id: int, db: Session = Depends(deps.get_db)):
-    s3.delete_audio_from_s3(db=db, audio_id=audio_id)
-    db.query(models.Audio).filter(models.Audio.audio_id == audio_id).delete()
-    db.commit()
-    return "Delete Successful"
+    return crud_audio.Audio.delete_audio(audio_id=audio_id,db=db)
