+++ backend/app/app/schemas/books.py
@@ -3,6 +3,7 @@ from fastapi import UploadFile
 
 class BookBase(BaseModel):
     book_name: str
+    book_file: str 
     # book_file: Optional[UploadFile] = None
 
 class BookCreate(BookBase):
@@ -12,9 +13,9 @@ class BookUpdate(BookBase):
     pass
 
 class BookInDBBase(BaseModel):
-    book_id: int
+    book_id: int   
     book_name: str
-    # book_file: Optional[str]  # Assuming you want to store file paths in the database
+    book_file: str  
 
     class Config:
         orm_mode = True
