+++ pyproject.toml
@@ -31,6 +31,7 @@ fitz = "^0.0.1.dev2"
 pdfplumber = "^0.10.3"
 aiofiles = "^23.2.1"
 boto3 = "^1.34.31"
+celery = "^5.3.6"
 
 [build-system]
 requires = ["poetry-core"]
